requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
Flask==3.1.3
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.9.4
mlflow==3.1.4
mlflow-skinny==3.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.2
pydantic_core==2.46.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.4.1
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.39.0
Werkzeug==3.1.8
zipp==3.23.1
